move to post-slice
