+++ packages/client/src/pages/home/map.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 import {
   GoogleMap,
   LoadScript,
@@ -7,7 +7,7 @@ import {
 } from "@react-google-maps/api";
 import { CSSProperties } from "@material-ui/styles";
 import { Libraries } from "@react-google-maps/api/dist/utils/make-load-script-url";
-import { Location } from "../../redux/slices/location-slice";
+import { Location } from "../../redux/slices/post-slice";
 import { setLocationFilter } from "../../redux/slices/post-slice";
 import REACT_APP_GOOGLE_API_KEY from "../../env";
 import styled from "styled-components";
@@ -88,9 +88,7 @@ const Map = () => {
 
   const [name, setName] = React.useState(location.name);
 
-  if (initialLocation === undefined) {
-    // TODO: this block is causing getPostsByFilter to be called twice when the page is refreshed
-    // Another issue: user location is fetched here AND in post-slice
+  useEffect(() => {
     navigator.geolocation.getCurrentPosition(function (position) {
       initialLocation = {
         lat: position.coords.latitude,
@@ -103,7 +101,7 @@ const Map = () => {
       });
       updateStore(initialLocation.lat, initialLocation.lon, location.name);
     });
-  }
+  }, [dispatch]);
 
   const updateOnDrag = (e: any) => {
     setCtr({ lat: e.latLng.lat(), lng: e.latLng.lng() });
